requirements.txt
Django==3.2.9
psycopg2-binary==2.9.1
djangorestframework==3.13.1
django-redis==5.0.0
celery==5.2.3